backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
pymongo>=4.5.0
mongoengine>=0.27.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0
pandas>=1.5.0